5758 - Move table hooks to their own file
