+++ src/client/pages/Section/DataTable/Table/DataValidations/DataValidations.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next'
 import { Table } from 'meta/assessment/table'
 
 import { useHistoryLastApprovedIsActive } from 'client/store/data/history/hooks/lastApproved'
-import { useTableHasErrors } from 'client/store/data/tableData/validations/hooks/validations'
+import { useTableHasErrors } from 'client/store/data/tableData/validations/hooks/tables'
 import { useCycle } from 'client/store/meta/hooks/cycles'
 import Icon from 'client/components/Icon'
