+++ src/client/pages/Section/DataTable/Table/RowData/Cell/Cell.tsx
@@ -11,7 +11,7 @@ import { TooltipId } from 'meta/tooltip/id'
 
 import { useHistoryLastApprovedIsActive } from 'client/store/data/history/hooks/lastApproved'
 import { useHistoryLastApprovedDataTableFetched } from 'client/store/data/history/hooks/lastApprovedTableData'
-import { useNodeValueValidation } from 'client/store/data/tableData/validations/hooks/validations'
+import { useNodeValueValidation } from 'client/store/data/tableData/validations/hooks/tables'
 import { useCycle } from 'client/store/meta/hooks/cycles'
 import { DataCell } from 'client/components/DataGrid'
 import { useCellId } from 'client/pages/Section/DataTable/Table/hooks/useCellId'
